requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
DateTime==5.3
idna==3.6
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
setuptools==69.0.2
urllib3==2.1.0
zope.interface==6.1
